[MarkScan] Start extracting PollWorkerScreen components for reuse
