+++ libs/mark-flow-ui/src/components/poll_worker/elements.tsx
@@ -1,0 +1,32 @@
+/* istanbul ignore file - @preserve - currently tested via apps. */
+
+import styled from 'styled-components';
+
+export const ButtonGrid = styled.div`
+  display: grid;
+  grid-auto-rows: 1fr;
+  grid-gap: max(${(p) => p.theme.sizes.minTouchAreaSeparationPx}px, 0.25rem);
+  grid-template-columns: 1fr 1fr;
+
+  button {
+    flex-wrap: nowrap;
+    white-space: nowrap;
+  }
+
+  margin-top: 0.5rem;
+`;
+
+export const VotingSession = styled.div`
+  margin: 30px 0 60px;
+  border: 2px solid #000;
+  border-radius: 20px;
+  padding: 30px 40px;
+
+  & > *:first-child {
+    margin-top: 0;
+  }
+
+  & > *:last-child {
+    margin-bottom: 0;
+  }
+`;
